+++ xmake/core/project/project.lua
@@ -163,7 +163,7 @@ function project._api_add_plugindirs(interp, ...)
     end
 
     -- add all plugins
-    interp:api_builtin_add_subdirs(plugindirs)
+    interp:api_builtin_includes(plugindirs)
 end
 
 -- add package directories and load all packages from the given directories
@@ -177,7 +177,7 @@ function project._api_add_packagedirs(interp, ...)
     end
 
     -- add all packages
-    interp:api_builtin_add_subdirs(pkgdirs)
+    interp:api_builtin_includes(pkgdirs)
 end
 
 -- get interpreter